Example input .csv file
This file shows the basic structure of how a .csv file should be formatted prior to conversion through the csv2ris script. In this example, the source metadata contained a column that contained useful information but did not conform to any of the standard ris tags. This column has been labelled "N1", so this information will go into the "Notes", "Comments", or "Extra" field in most reference managers. This is a useful way to preserve information that doesn't correspond well to a ris tag.